[MIG] stock_quant_cost_info: Migration to 15.0
- Module renamed from stock_inventory_cost_info to stock_quant_cost_info
- Add changes to allow show the cost difference when changing qty from stock.quants

TT36551
