+++ stock_quant_cost_info/README.rst
@@ -1,6 +1,6 @@
-=========================
-Stock Inventory Cost Info
-=========================
+=====================
+Stock Quant Cost Info
+=====================
 
 .. !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
    !! This file is generated by oca-gen-addon-readme !!
@@ -14,20 +14,20 @@ Stock Inventory Cost Info
     :target: http://www.gnu.org/licenses/agpl-3.0-standalone.html
     :alt: License: AGPL-3
 .. |badge3| image:: https://img.shields.io/badge/github-OCA%2Fstock--logistics--warehouse-lightgray.png?logo=github
-    :target: https://github.com/OCA/stock-logistics-warehouse/tree/14.0/stock_inventory_cost_info
+    :target: https://github.com/OCA/stock-logistics-warehouse/tree/15.0/stock_quant_cost_info
     :alt: OCA/stock-logistics-warehouse
 .. |badge4| image:: https://img.shields.io/badge/weblate-Translate%20me-F47D42.png
-    :target: https://translation.odoo-community.org/projects/stock-logistics-warehouse-14-0/stock-logistics-warehouse-14-0-stock_inventory_cost_info
+    :target: https://translation.odoo-community.org/projects/stock-logistics-warehouse-15-0/stock-logistics-warehouse-15-0-stock_quant_cost_info
     :alt: Translate me on Weblate
 .. |badge5| image:: https://img.shields.io/badge/runbot-Try%20me-875A7B.png
-    :target: https://runbot.odoo-community.org/runbot/153/14.0
+    :target: https://runbot.odoo-community.org/runbot/153/15.0
     :alt: Try me on Runbot
 
 |badge1| |badge2| |badge3| |badge4| |badge5| 
 
-This module extends the functionality of *Inventory Adjustments* to show a new
-column *Adjustment cost* in *Inventory Details* list and PDF report table. The
-value of *Adjustment cost* will be self-calculated according to:
+This module extends the functionality of *Quantity Adjustments* in *Stock Quants* to
+show a new column *Adjustment cost* in *Quant Details* list and PDF report table.
+The value of *Adjustment cost* will be self-calculated according to:
 (`Real Quantity` - `Theoretical Quantity`) * `Product Cost`
 
 **Table of contents**
@@ -39,18 +39,17 @@ Installation
 ============
 
 A pre_init_hook process is initiated in order to set the *Adjustment cost* to
-zero in all existing *inventory adjustment* before installation.
+zero in all existing *stock quant* before installation.
 
 Usage
 =====
 
 To use this module, you need to:
 
-#. Go to *Inventory > Operations > Inventory Adjustments* and create a new one.
-#. Click on *Start Inventory* button.
-#. In *Inventory Details* list you will see a new column named
-   *Adjustment cost*.
-#. The cost will be recomputed when the inventory adjustment is validated.
+#. Go to *Inventory > Products > Products* and create or select one.
+#. Click on *Update Quantity* button.
+#. In *Quant* list you will see a new column named *Adjustment cost*.
+#. The cost will be recomputed when we change the *Counted Quantity*.
 
 Bug Tracker
 ===========
@@ -58,7 +57,7 @@ Bug Tracker
 Bugs are tracked on `GitHub Issues <https://github.com/OCA/stock-logistics-warehouse/issues>`_.
 In case of trouble, please check there if your issue has already been reported.
 If you spotted it first, help us smashing it by providing a detailed and welcomed
-`feedback <https://github.com/OCA/stock-logistics-warehouse/issues/new?body=module:%20stock_inventory_cost_info%0Aversion:%2014.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
+`feedback <https://github.com/OCA/stock-logistics-warehouse/issues/new?body=module:%20stock_quant_cost_info%0Aversion:%2015.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
 
 Do not contact contributors directly about support or help with technical issues.
 
@@ -78,6 +77,7 @@ Contributors
   * Ernesto Tejeda
   * Pedro M. Baeza
   * Sergio Teruel
+  * Carlos Roca
 
 Maintainers
 ~~~~~~~~~~~
@@ -92,6 +92,6 @@ OCA, or the Odoo Community Association, is a nonprofit organization whose
 mission is to support the collaborative development of Odoo features and
 promote its widespread use.
 
-This module is part of the `OCA/stock-logistics-warehouse <https://github.com/OCA/stock-logistics-warehouse/tree/14.0/stock_inventory_cost_info>`_ project on GitHub.
+This module is part of the `OCA/stock-logistics-warehouse <https://github.com/OCA/stock-logistics-warehouse/tree/15.0/stock_quant_cost_info>`_ project on GitHub.
 
 You are welcome to contribute. To learn how please visit https://odoo-community.org/page/Contribute.
