+++ stock_quant_cost_info/hooks.py
@@ -4,12 +4,12 @@
 
 def pre_init_hook(cr):
     cr.execute(
-        """ALTER TABLE stock_inventory_line
+        """ALTER TABLE stock_quant
     ADD COLUMN adjustment_cost numeric
     DEFAULT 0"""
     )
 
     cr.execute(
-        """ALTER TABLE stock_inventory_line
+        """ALTER TABLE stock_quant
     ALTER COLUMN adjustment_cost DROP DEFAULT;"""
     )
