--- stock_quant_cost_info/i18n/stock_inventory_cost_info.pot
@@ -1,54 +1,0 @@
-# Translation of Odoo Server.
-# This file contains the translation of the following modules:
-# 	* stock_inventory_cost_info
-#
-msgid ""
-msgstr ""
-"Project-Id-Version: Odoo Server 14.0\n"
-"Report-Msgid-Bugs-To: \n"
-"Last-Translator: \n"
-"Language-Team: \n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: \n"
-"Plural-Forms: \n"
-
-#. module: stock_inventory_cost_info
-#: model_terms:ir.ui.view,arch_db:stock_inventory_cost_info.report_inventory_cost_info
-msgid "<strong>Adjustment cost</strong>"
-msgstr ""
-
-#. module: stock_inventory_cost_info
-#: model:ir.model.fields,field_description:stock_inventory_cost_info.field_stock_inventory_line__adjustment_cost
-msgid "Adjustment cost"
-msgstr ""
-
-#. module: stock_inventory_cost_info
-#: model:ir.model.fields,field_description:stock_inventory_cost_info.field_stock_inventory_line__currency_id
-msgid "Currency"
-msgstr ""
-
-#. module: stock_inventory_cost_info
-#: model:ir.model.fields,field_description:stock_inventory_cost_info.field_stock_inventory_line__display_name
-msgid "Display Name"
-msgstr ""
-
-#. module: stock_inventory_cost_info
-#: model:ir.model.fields,field_description:stock_inventory_cost_info.field_stock_inventory_line__id
-msgid "ID"
-msgstr ""
-
-#. module: stock_inventory_cost_info
-#: model:ir.model,name:stock_inventory_cost_info.model_stock_inventory_line
-msgid "Inventory Line"
-msgstr ""
-
-#. module: stock_inventory_cost_info
-#: model:ir.model.fields,field_description:stock_inventory_cost_info.field_stock_inventory_line____last_update
-msgid "Last Modified on"
-msgstr ""
-
-#. module: stock_inventory_cost_info
-#: model_terms:ir.ui.view,arch_db:stock_inventory_cost_info.stock_inventory_line_tree
-msgid "Total"
-msgstr ""
