+++ stock_quant_cost_info/i18n/stock_quant_cost_info.pot
@@ -1,46 +1,39 @@
 # Translation of Odoo Server.
 # This file contains the translation of the following modules:
-# 	* stock_inventory_cost_info
+# 	* stock_quant_cost_info
 #
 msgid ""
 msgstr ""
-"Project-Id-Version: Odoo Server 11.0\n"
+"Project-Id-Version: Odoo Server 15.0\n"
 "Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2023-01-16 09:20+0000\n"
-"PO-Revision-Date: 2023-01-16 10:21+0100\n"
-"Last-Translator: <>\n"
+"Last-Translator: \n"
 "Language-Team: \n"
-"Language: es\n"
 "MIME-Version: 1.0\n"
 "Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: 8bit\n"
+"Content-Transfer-Encoding: \n"
 "Plural-Forms: \n"
-"X-Generator: Poedit 3.2.2\n"
 
 #. module: stock_quant_cost_info
 #: model_terms:ir.ui.view,arch_db:stock_quant_cost_info.report_inventory_cost_info
 msgid "<strong>Adjustment cost</strong>"
-msgstr "<strong>Coste de ajuste</strong>"
+msgstr ""
 
 #. module: stock_quant_cost_info
 #: model:ir.model.fields,field_description:stock_quant_cost_info.field_stock_quant__adjustment_cost
 msgid "Adjustment cost"
-msgstr "Coste de ajuste"
+msgstr ""
 
 #. module: stock_quant_cost_info
 #: model:ir.model.fields,field_description:stock_quant_cost_info.field_stock_quant__currency_id
 msgid "Currency"
-msgstr "Moneda"
+msgstr ""
 
 #. module: stock_quant_cost_info
 #: model:ir.model,name:stock_quant_cost_info.model_stock_quant
 msgid "Quants"
-msgstr "Quants"
+msgstr ""
 
 #. module: stock_quant_cost_info
 #: model_terms:ir.ui.view,arch_db:stock_quant_cost_info.view_stock_quant_tree_inventory_editable
 msgid "Total"
-msgstr "Total"
-
-#~ msgid "Inventory Line"
-#~ msgstr "Línea de inventario"
+msgstr ""
