+++ stock_quant_cost_info/models/__init__.py
@@ -1,3 +1,3 @@
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl.html).
 
-from . import stock_inventory
+from . import stock_quant
