+++ stock_quant_cost_info/readme/CONTRIBUTORS.rst
@@ -3,3 +3,4 @@
   * Ernesto Tejeda
   * Pedro M. Baeza
   * Sergio Teruel
+  * Carlos Roca
