+++ stock_quant_cost_info/readme/DESCRIPTION.rst
@@ -1,4 +1,4 @@
-This module extends the functionality of *Inventory Adjustments* to show a new
-column *Adjustment cost* in *Inventory Details* list and PDF report table. The
-value of *Adjustment cost* will be self-calculated according to:
+This module extends the functionality of *Quantity Adjustments* in *Stock Quants* to
+show a new column *Adjustment cost* in *Quant Details* list and PDF report table.
+The value of *Adjustment cost* will be self-calculated according to:
 (`Real Quantity` - `Theoretical Quantity`) * `Product Cost`
