+++ stock_quant_cost_info/readme/INSTALL.rst
@@ -1,2 +1,2 @@
 A pre_init_hook process is initiated in order to set the *Adjustment cost* to
-zero in all existing *inventory adjustment* before installation.
+zero in all existing *stock quant* before installation.
