+++ stock_quant_cost_info/readme/USAGE.rst
@@ -1,7 +1,6 @@
 To use this module, you need to:
 
-#. Go to *Inventory > Operations > Inventory Adjustments* and create a new one.
-#. Click on *Start Inventory* button.
-#. In *Inventory Details* list you will see a new column named
-   *Adjustment cost*.
-#. The cost will be recomputed when the inventory adjustment is validated.
+#. Go to *Inventory > Products > Products* and create or select one.
+#. Click on *Update Quantity* button.
+#. In *Quant* list you will see a new column named *Adjustment cost*.
+#. The cost will be recomputed when we change the *Counted Quantity*.
