+++ stock_quant_cost_info/static/description/index.html
@@ -4,7 +4,7 @@
 <head>
 <meta http-equiv="Content-Type" content="text/html; charset=utf-8" />
 <meta name="generator" content="Docutils 0.15.1: http://docutils.sourceforge.net/" />
-<title>Stock Inventory Cost Info</title>
+<title>Stock Quant Cost Info</title>
 <style type="text/css">
 
 /*
@@ -360,17 +360,17 @@ ul.auto-toc {
 </style>
 </head>
 <body>
-<div class="document" id="stock-inventory-cost-info">
-<h1 class="title">Stock Inventory Cost Info</h1>
+<div class="document" id="stock-quant-cost-info">
+<h1 class="title">Stock Quant Cost Info</h1>
 
 <!-- !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
 !! This file is generated by oca-gen-addon-readme !!
 !! changes will be overwritten.                   !!
 !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! -->
-<p><a class="reference external" href="https://odoo-community.org/page/development-status"><img alt="Beta" src="https://img.shields.io/badge/maturity-Beta-yellow.png" /></a> <a class="reference external" href="http://www.gnu.org/licenses/agpl-3.0-standalone.html"><img alt="License: AGPL-3" src="https://img.shields.io/badge/licence-AGPL--3-blue.png" /></a> <a class="reference external" href="https://github.com/OCA/stock-logistics-warehouse/tree/14.0/stock_inventory_cost_info"><img alt="OCA/stock-logistics-warehouse" src="https://img.shields.io/badge/github-OCA%2Fstock--logistics--warehouse-lightgray.png?logo=github" /></a> <a class="reference external" href="https://translation.odoo-community.org/projects/stock-logistics-warehouse-14-0/stock-logistics-warehouse-14-0-stock_inventory_cost_info"><img alt="Translate me on Weblate" src="https://img.shields.io/badge/weblate-Translate%20me-F47D42.png" /></a> <a class="reference external" href="https://runbot.odoo-community.org/runbot/153/14.0"><img alt="Try me on Runbot" src="https://img.shields.io/badge/runbot-Try%20me-875A7B.png" /></a></p>
-<p>This module extends the functionality of <em>Inventory Adjustments</em> to show a new
-column <em>Adjustment cost</em> in <em>Inventory Details</em> list and PDF report table. The
-value of <em>Adjustment cost</em> will be self-calculated according to:
+<p><a class="reference external" href="https://odoo-community.org/page/development-status"><img alt="Beta" src="https://img.shields.io/badge/maturity-Beta-yellow.png" /></a> <a class="reference external" href="http://www.gnu.org/licenses/agpl-3.0-standalone.html"><img alt="License: AGPL-3" src="https://img.shields.io/badge/licence-AGPL--3-blue.png" /></a> <a class="reference external" href="https://github.com/OCA/stock-logistics-warehouse/tree/15.0/stock_quant_cost_info"><img alt="OCA/stock-logistics-warehouse" src="https://img.shields.io/badge/github-OCA%2Fstock--logistics--warehouse-lightgray.png?logo=github" /></a> <a class="reference external" href="https://translation.odoo-community.org/projects/stock-logistics-warehouse-15-0/stock-logistics-warehouse-15-0-stock_quant_cost_info"><img alt="Translate me on Weblate" src="https://img.shields.io/badge/weblate-Translate%20me-F47D42.png" /></a> <a class="reference external" href="https://runbot.odoo-community.org/runbot/153/15.0"><img alt="Try me on Runbot" src="https://img.shields.io/badge/runbot-Try%20me-875A7B.png" /></a></p>
+<p>This module extends the functionality of <em>Quantity Adjustments</em> in <em>Stock Quants</em> to
+show a new column <em>Adjustment cost</em> in <em>Quant Details</em> list and PDF report table.
+The value of <em>Adjustment cost</em> will be self-calculated according to:
 (<cite>Real Quantity</cite> - <cite>Theoretical Quantity</cite>) * <cite>Product Cost</cite></p>
 <p><strong>Table of contents</strong></p>
 <div class="contents local topic" id="contents">
@@ -389,17 +389,16 @@ value of <em>Adjustment cost</em> will be self-calculated according to:
 <div class="section" id="installation">
 <h1><a class="toc-backref" href="#id1">Installation</a></h1>
 <p>A pre_init_hook process is initiated in order to set the <em>Adjustment cost</em> to
-zero in all existing <em>inventory adjustment</em> before installation.</p>
+zero in all existing <em>stock quant</em> before installation.</p>
 </div>
 <div class="section" id="usage">
 <h1><a class="toc-backref" href="#id2">Usage</a></h1>
 <p>To use this module, you need to:</p>
 <ol class="arabic simple">
-<li>Go to <em>Inventory &gt; Operations &gt; Inventory Adjustments</em> and create a new one.</li>
-<li>Click on <em>Start Inventory</em> button.</li>
-<li>In <em>Inventory Details</em> list you will see a new column named
-<em>Adjustment cost</em>.</li>
-<li>The cost will be recomputed when the inventory adjustment is validated.</li>
+<li>Go to <em>Inventory &gt; Products &gt; Products</em> and create or select one.</li>
+<li>Click on <em>Update Quantity</em> button.</li>
+<li>In <em>Quant</em> list you will see a new column named <em>Adjustment cost</em>.</li>
+<li>The cost will be recomputed when we change the <em>Counted Quantity</em>.</li>
 </ol>
 </div>
 <div class="section" id="bug-tracker">
@@ -407,7 +406,7 @@ zero in all existing <em>inventory adjustment</em> before installation.</p>
 <p>Bugs are tracked on <a class="reference external" href="https://github.com/OCA/stock-logistics-warehouse/issues">GitHub Issues</a>.
 In case of trouble, please check there if your issue has already been reported.
 If you spotted it first, help us smashing it by providing a detailed and welcomed
-<a class="reference external" href="https://github.com/OCA/stock-logistics-warehouse/issues/new?body=module:%20stock_inventory_cost_info%0Aversion:%2014.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**">feedback</a>.</p>
+<a class="reference external" href="https://github.com/OCA/stock-logistics-warehouse/issues/new?body=module:%20stock_quant_cost_info%0Aversion:%2015.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**">feedback</a>.</p>
 <p>Do not contact contributors directly about support or help with technical issues.</p>
 </div>
 <div class="section" id="credits">
@@ -425,6 +424,7 @@ If you spotted it first, help us smashing it by providing a detailed and welcome
 <li>Ernesto Tejeda</li>
 <li>Pedro M. Baeza</li>
 <li>Sergio Teruel</li>
+<li>Carlos Roca</li>
 </ul>
 </li>
 </ul>
@@ -436,7 +436,7 @@ If you spotted it first, help us smashing it by providing a detailed and welcome
 <p>OCA, or the Odoo Community Association, is a nonprofit organization whose
 mission is to support the collaborative development of Odoo features and
 promote its widespread use.</p>
-<p>This module is part of the <a class="reference external" href="https://github.com/OCA/stock-logistics-warehouse/tree/14.0/stock_inventory_cost_info">OCA/stock-logistics-warehouse</a> project on GitHub.</p>
+<p>This module is part of the <a class="reference external" href="https://github.com/OCA/stock-logistics-warehouse/tree/15.0/stock_quant_cost_info">OCA/stock-logistics-warehouse</a> project on GitHub.</p>
 <p>You are welcome to contribute. To learn how please visit <a class="reference external" href="https://odoo-community.org/page/Contribute">https://odoo-community.org/page/Contribute</a>.</p>
 </div>
 </div>
