+++ stock_quant_cost_info/tests/__init__.py
@@ -1,3 +1,3 @@
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl.html).
 
-from . import test_stock_inventory_cost_info
+from . import test_stock_quant_cost_info
