+++ stock_quant_cost_info/views/stock_inventory_views.xml
@@ -1,13 +1,12 @@
 <?xml version="1.0" encoding="utf-8" ?>
 <odoo>
-    <record id="stock_inventory_line_tree" model="ir.ui.view">
-        <field name="name">stock.inventory.line.tree.cost.info</field>
-        <field name="model">stock.inventory.line</field>
-        <field name="inherit_id" ref="stock.stock_inventory_line_tree" />
+    <record id="view_stock_quant_tree_inventory_editable" model="ir.ui.view">
+        <field name="model">stock.quant</field>
+        <field name="inherit_id" ref="stock.view_stock_quant_tree_inventory_editable" />
         <field name="arch" type="xml">
-            <field name="product_uom_id" position="after">
+            <field name="inventory_diff_quantity" position="after">
                 <field name="currency_id" invisible="1" />
-                <field name="adjustment_cost" sum="Total" />
+                <field name="adjustment_cost" sum="Total" optional="show" />
             </field>
         </field>
     </record>
